Improve TraCI traffic light tutorial
