+++ docs/web/docs/Tutorials/TraCI4Traffic_Lights.md
@@ -42,44 +42,52 @@ python
 python runner.py
 ```
 
+The route file can also be regenerated independently by running
+`python generateRoutes.py` from the tutorial directory.
+
 !!! caution
     You need to press start in the simulation gui to run the tutorial.
 
 # Data preparation
 
-The net-definition can be found in the files `cross.nod.xml`,
-`cross.edg.xml`, `cross.con.xml` and `cross.det.xml`. The route data is
-generated randomly by the script. The vehicles leave the source
-according to a Poisson process approximated here by a binomial
-distribution. So a parameter `p=1./30` in the script means that a
-vehicle is generated every 30 seconds in average.
+The network definition is stored in `cross.nod.xml`, `cross.edg.xml`,
+`cross.con.xml` and `cross.det.xml`. The signalized intersection uses the
+ID `junction`, while the induction loop on the northern approach uses the
+ID `north_detector`.
+
+Vehicle demand is generated by `generateRoutes.py` and written to
+`data/cross.rou.xml`. The script uses a fixed random seed so that the
+example and its tests remain reproducible. Vehicle arrivals are sampled
+with a binomial approximation of a Poisson process. For example, a
+probability of `p=1./30` corresponds to one vehicle every 30 seconds on
+average.
 
 # Code
 
-The control logic resides in the python script *runner.py*. It generates
-the routes, acts with the server and controls the traffic light. It
-makes use of the TraCI python API bundled with SUMO. A description of
-the API can be found at
+The example separates scenario generation from online traffic control.
+`generateRoutes.py` creates the route file, whereas `runner.py` starts
+SUMO, connects through TraCI, reads the induction loop and controls the
+traffic light. The route generator is invoked by `runner.py`, so the full
+example can still be started with a single command.
+
+The controller uses the TraCI python API bundled with SUMO. A description
+of the API can be found at
 [TraCI/Interfacing_TraCI_from_Python](../TraCI/Interfacing_TraCI_from_Python.md).
 For a detailed list of available functions see the [pydoc generated
 documentation](https://sumo.dlr.de/pydoc/traci.html).
 
 # Simulation
 
-The main program is implemented in the script *runner.py*. There, we
-first generate the routes as described above. Then `traci.start` is used
-to start [sumo-gui](../sumo-gui.md) with the config file
-`cross.sumocfg`. The start call also connects our script with the
-[sumo-gui](../sumo-gui.md) running as a server.
+When `runner.py` starts, it first calls the route-generation function from
+`generateRoutes.py`. It then launches [sumo-gui](../sumo-gui.md) with
+`cross.sumocfg` by using `traci.start` and begins the control loop. The
+start call also connects the controller to the simulation server.
 
-Then we start to control the simulation. We let the server simulate one
-simulation step, read the induction loop and switch the phase of the
-traffic light until the end is reached where no vehicle exists on the
-road anymore. If we find a vehicle on the induction loop the phase is
-switched such that the north-south direction gets green. If no vehicle
-is on the detector, and we are not already in the process of switching
-(so EW has still green), we try to keep this phase by simply setting it
-again. In the end, we close the connection.
+During each simulation step, the controller checks `north_detector`. If a
+vehicle is detected while the east-west direction has green, the program
+switches the signal at `junction` toward the north-south phase. Otherwise,
+it keeps the current east-west phase. The TraCI connection is closed after
+no more vehicles are expected.
 
 # TraCI
