+++ tests/complex/tutorial/traci_tls/data/cross.det.xml
@@ -1,1 +1,1 @@
-<e1Detector id="0" lane="4i_0" pos="450" freq="30" file="cross.out" friendlyPos="x"/>
+<e1Detector id="north_detector" lane="4i_0" pos="450" freq="30" file="cross.out" friendlyPos="x"/>
