+++ tests/complex/tutorial/traci_tls/data/cross.edg.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <edges xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="http://sumo.dlr.de/xsd/edges_file.xsd">
-   <edge id="1i" from="1" to="0" priority="78" numLanes="1" speed="19.444" />
-   <edge id="1o" from="0" to="1" priority="46" numLanes="1" speed="11.111" />
+   <edge id="1i" from="1" to="junction" priority="78" numLanes="1" speed="19.444" />
+   <edge id="1o" from="junction" to="1" priority="46" numLanes="1" speed="11.111" />
 
-   <edge id="2i" from="2" to="0" priority="78" numLanes="1" speed="19.444" />
-   <edge id="2o" from="0" to="2" priority="46" numLanes="1" speed="11.111" />
+   <edge id="2i" from="2" to="junction" priority="78" numLanes="1" speed="19.444" />
+   <edge id="2o" from="junction" to="2" priority="46" numLanes="1" speed="11.111" />
 
-   <edge id="3i" from="3" to="0" priority="78" numLanes="1" speed="19.444" />
-   <edge id="3o" from="0" to="3" priority="46" numLanes="1" speed="11.111" />
+   <edge id="3i" from="3" to="junction" priority="78" numLanes="1" speed="19.444" />
+   <edge id="3o" from="junction" to="3" priority="46" numLanes="1" speed="11.111" />
 
-   <edge id="4i" from="4" to="0" priority="78" numLanes="1" speed="19.444" />
-   <edge id="4o" from="0" to="4" priority="46" numLanes="1" speed="11.111" />
+   <edge id="4i" from="4" to="junction" priority="78" numLanes="1" speed="19.444" />
+   <edge id="4o" from="junction" to="4" priority="46" numLanes="1" speed="11.111" />
 
    <edge id="51i" from="1" to="51" priority="78" numLanes="1" speed="19.444" />
    <edge id="51o" from="51" to="1" priority="46" numLanes="1" speed="11.111" />
