+++ tests/complex/tutorial/traci_tls/data/cross.nod.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <nodes xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="http://sumo.dlr.de/xsd/nodes_file.xsd">
-   <node id="0" x="0.0" y="0.0"  type="traffic_light"/>
+   <node id="junction" x="0.0" y="0.0"  type="traffic_light"/>
 
    <node id="1" x="-500.0" y="0.0" type="priority"/>
    <node id="2" x="+500.0" y="0.0" type="priority"/>
