+++ tests/complex/tutorial/traci_tls/generateRoutes.py
@@ -1,0 +1,62 @@
+#!/usr/bin/env python
+# Eclipse SUMO, Simulation of Urban MObility; see https://eclipse.dev/sumo
+# Copyright (C) 2009-2026 German Aerospace Center (DLR) and others.
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License 2.0 which is available at
+# https://www.eclipse.org/legal/epl-2.0/
+# This Source Code may also be made available under the following Secondary
+# Licenses when the conditions for such availability set forth in the Eclipse
+# Public License 2.0 are satisfied: GNU General Public License, version 2
+# or later which is available at
+# https://www.gnu.org/licenses/old-licenses/gpl-2.0-standalone.html
+# SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-or-later
+
+# @file    generateRoutes.py
+# @author  Lena Kalleske
+# @author  Daniel Krajzewicz
+# @author  Michael Behrisch
+# @author  Jakob Erdmann
+# @date    2009-03-26
+
+from __future__ import absolute_import
+from __future__ import print_function
+
+import random
+
+
+def generate_routefile():
+    """Generate deterministic route demand for the TraCI TLS tutorial."""
+    random.seed(42)  # make tests reproducible
+    N = 3600  # number of time steps
+    # demand per second from different directions
+    pWE = 1. / 10
+    pEW = 1. / 11
+    pNS = 1. / 30
+    with open("data/cross.rou.xml", "w") as routes:
+        print("""<routes>
+        <vType id="typeWE" accel="0.8" decel="4.5" sigma="0.5" length="5" minGap="2.5" maxSpeed="16.67" \
+guiShape="passenger"/>
+        <vType id="typeNS" accel="0.8" decel="4.5" sigma="0.5" length="7" minGap="3" maxSpeed="25" guiShape="bus"/>
+
+        <route id="right" edges="51o 1i 2o 52i" />
+        <route id="left" edges="52o 2i 1o 51i" />
+        <route id="down" edges="54o 4i 3o 53i" />""", file=routes)
+        vehNr = 0
+        for i in range(N):
+            if random.uniform(0, 1) < pWE:
+                print('    <vehicle id="right_%i" type="typeWE" route="right" depart="%i" />' % (
+                    vehNr, i), file=routes)
+                vehNr += 1
+            if random.uniform(0, 1) < pEW:
+                print('    <vehicle id="left_%i" type="typeWE" route="left" depart="%i" />' % (
+                    vehNr, i), file=routes)
+                vehNr += 1
+            if random.uniform(0, 1) < pNS:
+                print('    <vehicle id="down_%i" type="typeNS" route="down" depart="%i" color="1,0,0"/>' % (
+                    vehNr, i), file=routes)
+                vehNr += 1
+        print("</routes>", file=routes)
+
+
+if __name__ == "__main__":
+    generate_routefile()
