+++ tests/complex/tutorial/traci_tls/runner.py
@@ -24,7 +24,8 @@ from __future__ import print_function
 import os
 import sys
 import optparse
-import random
+
+from generateRoutes import generate_routefile
 
 # we need to import python modules from the $SUMO_HOME/tools directory
 if 'SUMO_HOME' in os.environ:
@@ -36,41 +37,11 @@ else:
 from sumolib import checkBinary  # noqa
 import traci  # noqa
 
-
-def generate_routefile():
-    random.seed(42)  # make tests reproducible
-    N = 3600  # number of time steps
-    # demand per second from different directions
-    pWE = 1. / 10
-    pEW = 1. / 11
-    pNS = 1. / 30
-    with open("data/cross.rou.xml", "w") as routes:
-        print("""<routes>
-        <vType id="typeWE" accel="0.8" decel="4.5" sigma="0.5" length="5" minGap="2.5" maxSpeed="16.67" \
-guiShape="passenger"/>
-        <vType id="typeNS" accel="0.8" decel="4.5" sigma="0.5" length="7" minGap="3" maxSpeed="25" guiShape="bus"/>
-
-        <route id="right" edges="51o 1i 2o 52i" />
-        <route id="left" edges="52o 2i 1o 51i" />
-        <route id="down" edges="54o 4i 3o 53i" />""", file=routes)
-        vehNr = 0
-        for i in range(N):
-            if random.uniform(0, 1) < pWE:
-                print('    <vehicle id="right_%i" type="typeWE" route="right" depart="%i" />' % (
-                    vehNr, i), file=routes)
-                vehNr += 1
-            if random.uniform(0, 1) < pEW:
-                print('    <vehicle id="left_%i" type="typeWE" route="left" depart="%i" />' % (
-                    vehNr, i), file=routes)
-                vehNr += 1
-            if random.uniform(0, 1) < pNS:
-                print('    <vehicle id="down_%i" type="typeNS" route="down" depart="%i" color="1,0,0"/>' % (
-                    vehNr, i), file=routes)
-                vehNr += 1
-        print("</routes>", file=routes)
+TLS_ID = "junction"
+DETECTOR_ID = "north_detector"
 
 # The program looks like this
-#    <tlLogic id="0" type="static" programID="0" offset="0">
+#    <tlLogic id="junction" type="static" programID="0" offset="0">
 # the locations of the tls are      NESW
 #        <phase duration="31" state="GrGr"/>
 #        <phase duration="6"  state="yryr"/>
@@ -83,17 +54,17 @@ def run():
     """execute the TraCI control loop"""
     step = 0
     # we start with phase 2 where EW has green
-    traci.trafficlight.setPhase("0", 2)
+    traci.trafficlight.setPhase(TLS_ID, 2)
     while traci.simulation.getMinExpectedNumber() > 0:
         traci.simulationStep()
-        if traci.trafficlight.getPhase("0") == 2:
+        if traci.trafficlight.getPhase(TLS_ID) == 2:
             # we are not already switching
-            if traci.inductionloop.getLastStepVehicleNumber("0") > 0:
+            if traci.inductionloop.getLastStepVehicleNumber(DETECTOR_ID) > 0:
                 # there is a vehicle from the north, switch
-                traci.trafficlight.setPhase("0", 3)
+                traci.trafficlight.setPhase(TLS_ID, 3)
             else:
                 # otherwise try to keep green for EW
-                traci.trafficlight.setPhase("0", 2)
+                traci.trafficlight.setPhase(TLS_ID, 2)
         step += 1
     traci.close()
     sys.stdout.flush()
